requirements.txt
click==8.0.3
python-dateutil==2.8.2
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
requests==2.22.0
Werkzeug==2.0.2

